+++ stream-chat-android-offline/src/main/java/io/getstream/chat/android/livedata/ChatDomain.kt
@@ -1,6 +1,0 @@
-package io.getstream.chat.android.livedata
-
-/**
- * The ChatDomain is the main entry point for all livedata & offline operations on chat.
- */
-public sealed interface ChatDomain